Refactoring Add Trait for common methods Add namespace Add entry method in main class
